Bill S469 version original - Luis R. Sepúlveda
BILL NUMBER: S469 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the criminal procedure law, in relation to the automatic expungement of arrest records PURPOSE: To automatically expunge arrest records when an individual was released without the filing of formal charges, a determination was made that the arrest was without probable cause or the proceeding against the individ- ual was dismissed. SUMMARY OF PROVISIONS: Section 1 adds a new section 160.61 to the criminal procedure law.