Bill S467 version original - Luis R. Sepúlveda
BILL NUMBER: S467 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to incarcerated individ- uals wages PURPOSE: To raise inmate wages for the first time in 25 years. SUMMARY OF PROVISIONS: Section 1 amends Section 187 of the correction law to raise incarcerated individuals wages to a minimum of one dollar and twenty cents per day and shall be given biannual raises for adequate performance in such assignments.