Bill S531 version original - Kevin Thomas
BILL NUMBER: S531 SPONSOR: THOMAS TITLE OF BILL: An act to amend the civil practice law and rules, in relation to the admissibility of statements in learned treatises, periodicals, or pamphlets PURPOSE OR GENERAL IDEA OF THE BILL: To adopt the Federal Rule of Evidence 803(18) regarding the hearsay exception for learned treatises. SUMMARY OF PROVISIONS: Section one of the bill amends the Civil Practice Law and Rules to add a new section 4549.