Bill S473 version original - Luis R. Sepúlveda
BILL NUMBER: S473 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the surrogate's court procedure act, in relation to court filing PURPOSE OR GENERAL IDEA OF BILL: The bill specifies grounds for a clerk to refuse to accept a filing SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new subdivision 4 to SCPA 2611. Section 2 is the effective date. This act take immediately.