Bill S474 version original - Luis R. Sepúlveda
BILL NUMBER: S474 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the executive law, in relation to appeals of parole determination PURPOSE: To change the parole appeal process to provide for timely and comprehen- sive review of parole denials. SUMMARY OF PROVISIONS: Section 1 amends subdivision 4 of section 259-i of the executive law. Section 2 amends subdivision 5 of section 259-i of the executive law.