Bill S501 version original - Dan Stec
BILL NUMBER: S501 SPONSOR: STEC TITLE OF BILL: An act to repeal sections of the highway law, relating to private roads PURPOSE: To eliminate this antiquated law as there are other mechanisms in place to address this issue. SUMMARY OF PROVISIONS: Section 1. Sections 300, 301, 302, 303, 304, 305, 306, 307, 308, 309, 310, 311, 312, 313, 314 and 315 of the highway law are repealed.