Bill S480 version original - Julia Salazar
BILL NUMBER: S480 SPONSOR: SALAZAR TITLE OF BILL: An act to amend the legislative law, in relation to creation of the non-partisan legislative budget office PURPOSE: Creates a nonpartisan budget office to provide budget related informa- tion to the committees of the senate and assembly. SUMMARY OF PROVISIONS: SECTION 1 of the bill creates a new Article 4-B of the Legislative Law, with new sections 76 and 77 as follows: Section 76. Non-partisan Legislative Budget Office. 1. Definitions. 2.