Bill S475 version original - Patrick M. Gallivan
BILL NUMBER: S475 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to the recertification proc- ess for licenses for firearms PURPOSE: The purpose of this bill is to eliminate the recertification requirement for firearms licenses created by the SAFE Act. SUMMARY OF PROVISIONS: Sections 1 amends the penal law, repealing the provisions that created a recertification requirement for firearms licenses. Section 2 states that this act shall take effect immediately.