Bill S494 version original - Dan Stec
BILL NUMBER: S494 SPONSOR: STEC TITLE OF BILL: An act to amend the penal law, in relation to providing that an elemen- tary or secondary school student shall be incapable of consenting to sexual conduct with a school employee PURPOSE: To criminalize school employees for sexual conduct with full time students of elementary and secondary schools. SUMMARY OF PROVISIONS: Section one amends section 130.00 of the penal law by defining "school employee.