Bill S507 version original - Kevin Thomas
BILL NUMBER: S507 SPONSOR: THOMAS TITLE OF BILL: An act to amend the insurance law, in relation to the establishment of the "Wellness Program Privacy Act" PURPOSE OR GENERAL IDEA OF BILL: This bill relates to the establishment of the Wellness Program Privacy Act. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill establishes that this act shall be known and may be cited as the "Wellness Program Privacy Act.