Bill S502 version original - Gustavo Rivera
BILL NUMBER: S502 SPONSOR: RIVERA TITLE OF BILL: An act to amend the criminal procedure law, in relation to the expunge- ment of syringe convictions PURPOSE: To expunge conviction records related to the possession or sale of hypo- dermic syringes and needles. SUMMARY OF PROVISIONS: Section 1 amends section 160.50 of the criminal procedure law to add a new subparagraph vii that provides for expungement of a conviction for an offense related to criminally possessing a hypodermic instrument.