Bill S497 version original - Dan Stec
BILL NUMBER: S497 SPONSOR: STEC TITLE OF BILL: An act to amend the general municipal law, in relation to allowing an authorized organization to receive a license to hold a game of chance on Easter Sunday and Christmas Day PURPOSE: Allows authorized organizations to hold a game of chance on Easter Sunday and Christmas Day. SUMMARY OF PROVISIONS: Section 1 amends the general municipal law in order for games of chance to be conducted on Easter Sunday or Christmas Day.