Bill S571 version original - Brian Kavanagh
BILL NUMBER: S571 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the legislative law, in relation to restrictions on political contributions from lobbyists PURPOSE: The bill limits the dollar amount of political contributions that lobby- ists can give to members of the legislature and that members can receive from lobbyists, and when the contributions can be given. SUMMARY OF PROVISIONS: The bill adds new sections 1-u and 1-v to the legislative law.