Bill S593A version original - Robert Ortt
BILL NUMBER: S593 SPONSOR: ORTT TITLE OF BILL: An act to amend the mental hygiene law, in relation to Down syndrome diagnosis awareness PURPOSE: To require information on Down syndrome be provided to a pregnant woman, parent, or expectant parent of any infant or fetus diagnosed with Down syndrome. SUMMARY OF PROVISIONS: Add a new section 13.