Bill S649 version original - Robert Ortt
BILL NUMBER: S649 SPONSOR: ORTT TITLE OF BILL: An act to amend the public health law, in relation to the "Unpredictable Nursing Home Inspection Act" PURPOSE: To require that the department of health conducts nursing home inspections on nights, weekends, and holidays. SUMMARY OF PROVISIONS: Section 1 allows the bill to be referred to as the "Unpredictable Nurs- ing Home Inspection Act." Section 2 defines "business hours" and "non-business hours.