Bill S670 version original - James Tedisco
BILL NUMBER: S670 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to taxable estates; and to repeal article 26 of the tax law relating thereto PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to repeal New York State's estate tax. SUMMARY OF SPECIFIC PROVISIONS: Section one repeals Article 26 of the tax law. § 2.