Bill S618 version original - Brian Kavanagh
BILL NUMBER: S618 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the state finance law, in relation to the definition and purchase of tropical hardwoods PURPOSE: To require that tropical hardwoods or tropical hardwood products that the state purchases are from sustainably managed forests. SUMMARY OF PROVISIONS: Section 1 of the bill amends subdivision 1 of section 165 of the state finance law, adding six new paragraphs that set forth legislative defi- nitions.