Bill S605A version original - Brian Kavanagh
BILL NUMBER: S605 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the release of certain balloons PURPOSE: Prohibits the mass release of helium balloons into the environment. SUMMARY OF PROVISIONS: Section 1 of the bill outlines the legislative findings. The release into the atmosphere of large numbers of helium balloons poses a danger to the environment, particularly to wildlife and marine animals, and results in litter.