Bill S656 version original - Robert Ortt
BILL NUMBER: S656 SPONSOR: ORTT TITLE OF BILL: An act to amend the agriculture and markets law, in relation to local laws restricting farming operations within an agricultural district PURPOSE: To solidify protections afforded to farmers under the Right to Farm Law. SUMMARY OF PROVISIONS: Section 1 amends the agriculture and markets law to remove the 250 acres minimum for the creation of an agricultural district.