Bill S589 version original - Liz Krueger
BILL NUMBER: S589 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the environmental conservation law, in relation to the prohibition on canned shoots PURPOSE OR GENERAL IDEA OF BILL: To prohibit the killing, shooting, maiming or injuring of certain inten- tionally confined animals for sport, amusement or taking of a trophy. SUMMARY OF SPECIFIC PROVISIONS: Paragraph a of subdivision 1 of section 11-1904 of the environmental conservation law, as added by chapter 208 of the laws of 1999 is amended.