Bill S669 version original - James Tedisco
BILL NUMBER: S669 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the correction law, in relation to enacting the "prison privilege limitation act" PURPOSE: To enact the "Prison Privilege Limitation Act." SUMMARY OF PROVISIONS: The correction law is amended by adding a new section 627 to read as follows: § 627. Limitation of prison privileges for certain individuals.