Bill S611B version original - Shelley Mayer
BILL NUMBER: S611 SPONSOR: MAYER TITLE OF BILL: An act to amend the election law, in relation to making commissioners full time employees of the board PURPOSE: To end the actual and perceived conflicts of interests of board of elections employees caused by outside income related to the adminis- tration of elections and supervision of elections in which they are candidates. SUMMARY OF SPECIFIC PROVISIONS: Section 1.