Bill S658 version original - Robert Ortt
BILL NUMBER: S658 SPONSOR: ORTT TITLE OF BILL: An act to amend the election law, in relation to establishing no school shall serve as a polling place PURPOSE: This legislation would expressly prevent any elementary or secondary school from being used as a poll site for any election. SUMMARY OF PROVISIONS: Amends election law to state that no polling place shall be located on the premise of any public or private elementary or secondary school or a school for preschool children.