Bill S596 version original - Robert Ortt
BILL NUMBER: S596 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to removing the maximum age limit for firearms instruction exemptions PURPOSE: Would allow any individual over the age of 12 to receive firearm instruction. SUMMARY OF PROVISIONS: Section 1 of this bill amends paragraph 7 of subdivision (a) of section 265.20 of the penal law to remove language limiting firearm instruction to individuals between the ages of 12 and 16.