Bill S665 version original - Brian Kavanagh
BILL NUMBER: S665 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to the phase-out of state use of pesticides PURPOSE: To implement a policy that will discontinue the use of pesticides by the state and adopt a pest control policy that substantially relies on non- chemical pest control policies. SUMMARY OF PROVISIONS: This bill will create a three-level phase-out of pesticide use by agen- cies of the State of New York.