Bill S646 version original - Brian Kavanagh
BILL NUMBER: S646 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the public authorities law, in relation to the use of wood in subway and railroad tracks PURPOSE: To require that tropical hardwoods or tropical hardwood products that the state purchases are from sustainably managed forests. SUMMARY OF PROVISIONS: Section 1 of the bill amends Section 1209 of the public authorities law prohibiting the procurement of tropical hardwoods for new or replacement subway and railroad track ties.