Bill S620 version original - Liz Krueger
BILL NUMBER: S620 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the executive law, in relation to cost benefit analysis of tax expenditures PURPOSE: This legislation proposes to require the Governor to provide in the annual Tax Expenditure Report more information -- including cost/benefit analyses that better gauges the effectiveness and efficiency of each tax expenditure. SUMMARY OF PROVISIONS: § 1.