Bill S671 version original - James Tedisco
BILL NUMBER: S671 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the penal law, in relation to establishing the bystander protection act PURPOSE: This Legislation will create additional penalties and fines for those or killing an innocent bystander during an attack convicted of injuring on an elected official. SUMMARY OF PROVISIONS: Section 1: Establishes the title of this bill as the "Bystander Protection Act." Section 2: Amends penal law by adding section 55.15.