Bill S585A version original - Rachel May
BILL NUMBER: S585 SPONSOR: MAY TITLE OF BILL: An act to amend the election law, in relation to the removal of an election commissioner PURPOSE: To ensure greater independence and accountability in the administration of New York's elections. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Amends subdivision 7 of section 3-200 of the Election Law. Section 2: sets the effective date. JUSTIFICATION: New York State's elections are administered at the county level by bipartisan commissioners.