Bill S659 version original - Robert Ortt
BILL NUMBER: S659 SPONSOR: ORTT TITLE OF BILL: An act to amend the canal law, in relation to maintaining bridges in a manner to not impede commercial motor vehicles PURPOSE: This legislation would require the State to maintain canal lift bridges in a manner not to impede commercial vehicle traffic as defined in subdivision four-a of section two of the transportation law. Further this legislation requires at least on public hearing in a community where the bridge to be closed is located.