Bill S595 version original - Robert Ortt
BILL NUMBER: S595 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to the criminalization of selling, using or possessing synthetic drugs PURPOSE: This bill would add synthetic hallucinogenic compounds and chemical derivatives on the controlled substance list. SUMMARY OF PROVISIONS: Section 1 Amend the penal law adding a new section 220.66 to criminalize the sale, use of synthetic drugs and other similar compounds. Section 2 is the severability clause.