Bill S609 version original - Brad Hoylman-Sigal
BILL NUMBER: S609 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the election law, in relation to split shifts for election inspectors or poll clerks PURPOSE OR GENERAL IDEA OF BILL: To allow poll clerks to work half-day shifts on Election Day. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 7 of § 3-400 of the Election Law to allow for half-day Election Day shifts for poll clerks. Section 2 contains the effective date.