Bill S641 version original - Brian Kavanagh
BILL NUMBER: S641 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the general obligations law and the labor law, in relation to the duties of skiers and snowboarders PURPOSE: To amend the general obligations law and the labor law, by requiring that children under the age of eighteen years wear protective headgear while skiing and snowboarding. PROVISIONS: Section 1 of the bill amends section 18-1.02 of the general obligations law by adding a new subdivision 8 defining "snowboarder".