Bill S653 version original - Robert Ortt
BILL NUMBER: S653 SPONSOR: ORTT TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to distinctive plates for medal of honor recipients PURPOSE OF THE BILL: Corrects an error on distinctive plates for medal of honor recipients. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 404-p of the Vehicle and Traffic Law to provide for distinctive plates for medal of honor recipients.