Bill S583 version original - Brad Hoylman-Sigal
BILL NUMBER: S583 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the environmental conservation law, in relation to enacting the dark skies protection act PURPOSE: SUMMARY OF SPECIFIC PROVISION: Section 1 of the bill the bill the "Dark Skies Act." Section 2 of the bill adds a new article 18 to the environmental conser- vation for the control of light pollution in New limits on the use of outdoor lighting. Section 18-0101 of Article Skies Protection Act.