Bill S696 version original - Robert Ortt
BILL NUMBER: S696 SPONSOR: ORTT TITLE OF BILL: An act to amend the environmental conservation law and the penal law, in relation to removing prohibition on possession and use of firearm silencers PURPOSE: This legislation would allow for the legal possession of firearm suppressors. SUMMARY OF PROVISIONS: Section 1 allows for the use of firearm suppressors for the use of hunt- ing. Section 2 allows for the possession of a firearm suppressor.