Bill S752 version original - Luis R. Sepúlveda
BILL NUMBER: S752 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the labor law, in relation to project labor agreement feasibility studies and apprenticeship agreements PURPOSE: To require a feasibility study prior to engagement in any project labor and apprenticeship agreement. SUMMARY OF PROVISIONS: Section 1 amends subdivision 2 of section 222 of the labor law. Section 2 amends section 2 of section 816-b of the labor law, as added by chapter 571 of the laws of 2001.