Bill S788 version original - Robert Ortt
BILL NUMBER: S788 SPONSOR: ORTT TITLE OF BILL: An act to amend the labor law, in relation to establishing a leave of absence for veterans PURPOSE: This legislation would allow a Veteran to have off for Veterans Day. SUMMARY OF SPECIFIC PROVISIONS: Section 1 creates a new section 202-n of labor law to allow a veteran to be granted a leave of absence from their employer on Veterans Day if they qualify as a veteran. This section does allow for a hardship waiver for business.