Bill S849 version original - Roxanne J. Persaud
BILL NUMBER: S849 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the executive law, in relation to requirements for in-service training programs for code enforcement personnel PURPOSE: This is a chapter amendment that makes changes to provisions L.2022 c.797, to remove the increased number of training hours required. SUMMARY OF PROVISIONS: This legislation amends the underlying chapter by removing provisions that increased the number of training hours required. EXISTING LAW: None.