Bill S767 version original - Jeremy Cooney
BILL NUMBER: S767 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to operation of an all-terrain vehicle by a minor on public or private property PURPOSE: To ensure safety for operators of all-terrain vehicles on public or private property. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 2410 of the vehicle and traffic law, which relates to the operation of all-terrain vehicles by minors.