Bill S682A version original - Leroy Comrie
BILL NUMBER: S682 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general business law, in relation to methods for cancelling contracts PURPOSE: To allow consumers to cancel a contract in the same way that they entered such contract. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 391-u of the general business law by adding a new section, which reads: Method for contract cancellation.