Bill S807 version original - José M. Serrano
BILL NUMBER: S807 SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to prohibiting the possession of concealed firearms in national parks and national wildlife refuge systems PURPOSE OR GENERAL IDEA OF THE BILL: To prohibit the possession of concealed firearms in any park, campground or other recreation areas. SUMMARY OF PROVISIONS: The parks, recreation and historical preservation law is amended by adding a new section 13.