Bill S683A version original - Leroy Comrie
BILL NUMBER: S683 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general business law, in relation to changes in energy service pricing PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to protect consumers from unknown price increases to their energy bills. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Would declare a change in the price or type of price for energy service to be a "material" change which would require the consent of the customer.