Bill S813 version original - John Liu
BILL NUMBER: S813 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to dates of commencement of membership of certain positions of school governance in the city of New York PURPOSE OF BILL: This is a chapter amendment that makes changes to provisions of L.2022, c.364. SUMMARY OF SPECIFIC PROVISIONS: This legislation amends the underlying chapter by extending mayoral control an additional two years and provides more time for the expanded PEP to be operational.