Bill S835 version original - Leroy Comrie
BILL NUMBER: S835 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public health law, in relation to guidelines for onsite opioid overdose response capacity in nightlife establishments; and to repeal section 3309-b of such law relating to the nightlife opioid antagonist program PURPOSE: This is a chapter amendment that makes changes to provisions of L.2022, c. , relating to establishing the nightlife opioid antagonist program.