Bill S695 version original - Robert Ortt
BILL NUMBER: S695 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to the criminal possession of fentanyl PURPOSE: The purpose of this legislation is so criminalize the unlawful possession of fentanyl or fentanyl analog. SUMMARY OF SPECIFIC PROVISIONS: The bill adds a new section 220.