Bill S690 version original - Robert Ortt
BILL NUMBER: S690 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to authorizing the sheriff of any county to issue pistol permits PURPOSE: This legislation allows the sheriff of any county outside of NYC to issue a license to carry or possess a pistol or revolver under section 400.00 if they so elect. SUMMARY OF SPECIFIC PROVISIONS: Section 1 allows the sheriff of any county outside of NYC to issue a license to carry or possess a pistol or revolver under section 400.