Bill S862 version original - Robert Ortt
BILL NUMBER: S862 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to sentences of intermittent imprisonment for sexual offenses PURPOSE: This legislation would prohibit an individual convicted of a sex offense from being eligible for intermittent sentences. SUMMARY OF PROVISIONS: Section 1 amends Subdivision 2 of section 85.00 of the penal law as it relates to intermittent imprisonment.