Bill S847 version original - Michelle Hinchey
BILL NUMBER: S847 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to natural fiber textile development PURPOSE: This is a chapter amendment that makes changes to provisions of L.2022, c.572, by replacing the Natural fiber and textile development council with the Natural fiber and textile development workgroup. This change will eliminate redundant work by agencies.