Bill S869 version original - Robert Ortt
BILL NUMBER: S869 SPONSOR: ORTT TITLE OF BILL: An act to amend the state administrative procedure act, in relation to enacting the "red tape reduction act" PURPOSE: This legislation would require that when a rule is adopted that imposes a new administrative burden on a business, one or more existing rules must be amended or repealed to offset the cost of the new administrative burden. Also requires the identification of the rule to be repealed whenever a new rule is proposed.