Bill S806 version original - Robert Ortt
BILL NUMBER: S806 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to providing a tax credit to farmers who purchase biodiesel fuel for the operation of their farm equipment PURPOSE: To create incentives for farmers to use homegrown biodiesel for opera- tion of their farm equipment rather than petrogas. SUMMARY OF SPECIFIC PROVISIONS: Section 1 and 2: Establishes a tax credit of $0.