Bill S769B version original - Jeremy Cooney
BILL NUMBER: S769 SPONSOR: COONEY TITLE OF BILL: An act to amend the education law, in relation to creating the profes- sion of nurse anesthetist PURPOSE: This bill creates a scope of practice for certified registered nurse anesthetists SUMMARY OF PROVISIONS: Section 1 of this bill creates a new Article 139-A of the education law and defines the scope of practice for certified registered nurse anesth- etist. Section 2 amends the references to who can use the title of nurse.