Bill S779 version original - Robert Ortt
BILL NUMBER: S779 SPONSOR: ORTT TITLE OF BILL: An act to amend the mental hygiene law, in relation to fairness in organ donations to persons with physical or mental disabilities PURPOSE: To ensure that individuals with a physical or mental disability are not discriminated against in organ donation. SUMMARY OF PROVISIONS: Section 1 adds a new article 35 to the mental hygiene law creating the organ donation fairness act.