Bill S726 version original - José M. Serrano
BILL NUMBER: S726 SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to providing for a review process of proposals to alienate municipal parkland PURPOSE OR GENERAL IDEA OF BILL: Municipally owned parkland is vital to local communities. Once lost, municipal parkland is difficult to recover.