Bill S748 version original - Luis R. Sepúlveda
BILL NUMBER: S748 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the surrogate's court procedure act, in relation to prohibiting the use of a waiver of probate form PURPOSE OR GENERAL IDEA OF BILL: This bill would prohibit the usage of a waiver of probate to transfer an inheritance, and prohibit a transfer agent from accepting a waiver of probate on behalf of a New York decedent. SUMMARY OF PROVISIONS: Section 1: Amends the Surrogate's Court Procedure Act by adding a new section 1409-a.