Bill S809A version original - José M. Serrano
BILL NUMBER: S809 SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to prohibiting littering on state park lands PURPOSE: To create a greater deterrent to littering in State Parks and Historic Sites. SUMMARY OF SPECIFIC PROVISIONS: -Codifies Selection 375.1 of the Office of Parks, Recreation and Histor- ic Preservation Codes, Rules and Regulations by prohibiting the throwing of refuse on park lands.