Bill S681 version original - George Borrello
BILL NUMBER: S681 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the state finance law, in relation to establishing a spending cap PURPOSE: This legislation will cap the growth of state spending and increase the allowable reserve of the rainy day fund. SUMMARY OF PROVISIONS: Section one of the bill would amend the State Finance Law ("SFL") to add a new Article 17, establishing a state spending cap. * New SFL § 244 sets for definitions for article 17.