Bill S868 version original - Robert Ortt
BILL NUMBER: S868 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to gun license applications not containing warnings PURPOSE: To prevent municipalities from the distributing of literature containing explicit warnings regarding the ownership or use of a firearm. SUMMARY OF PROVISIONS: Section 1 amends Paragraph (a) of subdivision 3 of section 400.