Bill S734 version original - Kevin S. Parker
BILL NUMBER: S734 SPONSOR: PARKER TITLE OF BILL: An act to amend the environmental conservation law, in relation to envi- ronmental benefit projects in lieu of civil penalties PURPOSE OR GENERAL IDEA OF THE BILL: The bill codifies requirements in the environmental conservation law in relation to requirements for partial settlements in the form of environ- mental benefit projects. Most of the requirements are currently contained in a draft Departmental guidance document.