Bill S837 version original - Gustavo Rivera
BILL NUMBER: S837 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to protecting the confidentiality of vaccine information; and to repeal provisions of the public health law relating thereto PURPOSE: This is a chapter amendment that makes changes to provisions of L.2022, c. , relating to protecting the confidentiality of vaccine information.