Bill S870 version original - Robert Ortt
BILL NUMBER: S870 SPONSOR: ORTT TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to the constitution, in relation to the right to hunt, trap and fish SUMMARY OF PROVISIONS: A new Article Twenty would be added to the state constitution to ensure the right of the law-abiding citizen to hunt, trap and fish subject to reasonable regulation by the state. JUSTIFICATION: The roots of hunting, fishing and trapping go deep into the soil of our American experience.