Bill S829 version original - James Sanders Jr.
BILL NUMBER: S829 SPONSOR: SANDERS TITLE OF BILL: An act to amend the agriculture and markets law, in relation to the establishment and priorities of the New York state council on hunger and food policy PURPOSE OF BILL: This is a chapter amendment that makes changes to provisions of L.2022, c.645. SUMMARY OF SPECIFIC PROVISIONS: This legislation amends the underlying chapter by expanding the purpose of the council to include hunger along with food policy.