Bill S878 version original - John Liu
BILL NUMBER: S878 SPONSOR: LIU TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the adjudi- cation of parking infractions PURPOSE: The bill would ensure that individuals who are issued a ticket for park- ing infractions receive a fair and sound process towards resolution. SUMMARY OF PROVISIONS: Section 1 provides that the New York state mandatory surcharge is always collected in instances where a notice of violation is not dismissed.