Bill S738 version original - Kevin S. Parker
BILL NUMBER: S738 SPONSOR: PARKER TITLE OF BILL: An act in relation to renaming the Newkirk Avenue subway station on the IRT Nostrand Avenue line the "Newkirk Avenue - Little Haiti station" PURPOSE: This bill will rename the current "Newkirk Ave" station on the 2 and 5 line to "Newkirk Avenue -- Little Haiti". The renaming will enhance the experience of the Haitian culture to visitors and other residence. SUMMARY OF PROVISIONS: Section 1.