Bill S801 version original - José M. Serrano
BILL NUMBER: S801 SPONSOR: SERRANO TITLE OF BILL: An act to amend the multiple dwelling law, in relation to rent reduction in cases of submetered utility service PURPOSE: The bill would require a rent reduction equal to the amount of the addi- tional. charge for the utilities for tenants in buildings which convert to submetering of utilities.