Bill S811 version original - Neil D. Breslin
BILL NUMBER: S811 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to annuity benefits PURPOSE: This is a chapter amendment that makes changes to provisions of L.2022, c., to clarify provisions relating to annuity benefits SUMMARY OF PROVISIONS: The changes are technical to conform to bill drafting nomenclature. EXISTING LAW: None JUSTIFICATION: This legislation is a negotiated change to the underlying chapter.