Bill S692 version original - Robert Ortt
BILL NUMBER: S692 SPONSOR: ORTT TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to providing veterans free use of cabins and cottages located in public places of recreation within the state PURPOSE: This legislation allows for the free use of cabins and cottages for any blind, non-ambulatory, or amputee or any veteran with certain service based disabilities for a period of fourteen days in a year. SUMMARY OF PROVISIONS: Section 1 amends Section 13.