Bill S731 version original - Kevin S. Parker
BILL NUMBER: S731 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to quality standards for internet service providers PURPOSE: To establish service quality standards for the provision of internet service. SUMMARY OF SPECIFIC PROVISIONS: Section 1 - The heading of article 11 of the public service law, as added by chapter 83 of the laws of 1995 is amended to include internet service providers (ISPs).