Bill S867 version original - Robert Ortt
BILL NUMBER: S867 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to establishing a credit against income tax for service dogs PURPOSE: To provide a tax credit to owners of service dogs. SUMMARY OF PROVISIONS: Section I: Establishes a tax credit of up to one thousand dollars for owners of qualified services dogs.