Bill S721 version original - José M. Serrano
BILL NUMBER: S721 SPONSOR: SERRANO TITLE OF BILL: An act to create a temporary state commission to study and make recom- mendations concerning the incidence of asthma in the borough of the Bronx in New York city PURPOSE OR GENERAL IDEA OF BILL: To create a state commission to study the incidence of asthma in Bronx County, the county with the highest incidence of asthma in the state. of New York.