Bill S887 version original - Cordell Cleare
BILL NUMBER: S887 SPONSOR: CLEARE TITLE OF BILL: An act to amend the elder law, in relation to establishing a dementia and Alzheimer's disease program database PURPOSE OR GENERAL IDEA OF BILL: This bill would encourage the sharing of information about programs that treat and support elderly individuals with Alzheimer's disease or demen- tia. SUMMARY OF SPECIFIC PROVISIONS: Section 1.