Bill S895B version original - Brad Hoylman-Sigal
BILL NUMBER: S895 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the general business law, in relation to requiring disclosure of certain social media terms of service PURPOSE: Requires social media companies, as defined, to post their terms of service and to submit reports to the Attorney General on their terms of service and content moderation policies and outcomes. SUMMARY OF PROVISIONS: Section one of the bill amends the general business law by adding a new article 42.