Bill S971 version original - Jamaal Bailey
BILL NUMBER: S971 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to confidential informants PURPOSE OR GENERAL IDEA OF BILL: To amend the law in relation to confidential informants by defining confidential informants; by imposing restrictions on who and who may not be a confidential informant; by regulating the use of confidential informants. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Adds a new section 60.23 to the criminal procedure law. 60.23.