Bill S947 version original - Kevin Thomas
BILL NUMBER: S947 SPONSOR: THOMAS TITLE OF BILL: An act to amend the penal law, in relation to unauthorized use of a smart phone or mobile device or any data contained therein PURPOSE OR GENERAL IDEA OF BILL: To protect personal information contained in cell phones in the same manner as personal information derived from computers. SUMMARY OF PROVISIONS: Section one amends paragraph one of § 156.00 of the penal law by adding smart phones and mobile devices to the definition of a computer.