Bill S933 version original - Roxanne J. Persaud
BILL NUMBER: S933 SPONSOR: PERSAUD TITLE OF BILL: An act relating to enacting the "Endoscope Reform Act" PURPOSE: The purpose of this bill is to promulgate rules and regulations to govern the practice of all upper endoscopic procedures. Currently, the vast majority of upper endoscopic procedures are being performed through a method involving sedation, which carries significant health risks and great financial cost.