Bill S1056A version original - Michelle Hinchey
BILL NUMBER: S1056 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to land used in agricultural production PURPOSE: This bill includes field crops grown and consumed by animals on a farm for the purpose of defining land used in agricultural production. SUMMARY OF PROVISIONS:. This bill removes the exclusion of field crops grown and consumed on a farm from the calculation of farm income for the purpose of eligibility of the definition of a farm.