Bill S907 version original - James Tedisco
BILL NUMBER: S907 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to estab- lishing a public registry for persons convicted of violating "Buster's Law" PURPOSE: To require each violator of "Buster's Law" to register his or her name and address with the division of criminal justice services. SUMMARY OF PROVISIONS: Section one - amend the agriculture and markets law by adding a new section 352 to read as follows: Animal cruelty register.