Bill S988 version original - Robert Ortt
BILL NUMBER: S988 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to the definition of person PURPOSE: This legislation names the provision related to person of Subdivision 1 of Section 125.05 of the penal law as Mya's Law. SUMMARY OF SPECIFIC PROVISIONS: Section 1 This legislation names the provision related to person of Subdivision 1 of Section125.05 of the penal law as Mya's Law. Section 2 is the effective date.