Bill S977 version original - Mario Mattera
BILL NUMBER: S977 SPONSOR: MATTERA TITLE OF BILL: An act to amend the criminal procedure law, in relation to authorizing bail for certain offenses committed during a riot PURPOSE OR GENERAL IDEA OF BILL: To categorize certain offenses committed during a riot as bail eligible. SUMMARY OF PROVISIONS: Section 1: amends paragraphs (a) and (t) of subdivision 4 of section 510.10 of the criminal procedure law and adds a new paragraph (u).