Bill S914 version original - John Liu
BILL NUMBER: S914 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to requiring a course of study to teach the history of Yu Gwan-Sun and the March 1st movement of Korea PURPOSE: This legislation aims to establish Yu Gwan-Sun's Law, which provides for the instruction of the life and work of. Korean activist Yu Gwan-Sun to all public school students in grades nine to twelve. SUMMARY OF PROVISIONS: Section 1 provides the short title of this Act as "Yu-Gwan-Sun's Law.